requirements.txt
flask==3.0.3
ultralytics>=8.3.0
torch>=2.4.0
torchvision>=0.19.0
numpy>=1.26.4,<2.0
pillow>=10.0.0
scipy>=1.11.0
python-dotenv>=1.0.1
opencv-python>=4.9.0.80
scikit-learn>=1.3.0